drm/compositor: Refactor buffer handling
